Break aggregation into classes for easier reading.
