+++ std-bits/table/src/main/java/org/enso/table/data/table/Table.java
@@ -47,7 +47,7 @@ public class Table {
   private Table(Column[] columns, Index index, AggregatedProblems problems) {
     this.columns = columns;
     this.index = index == null ? (new DefaultIndex((columns == null || columns.length == 0) ? 0 : columns[0].getSize())) : index;
-    this.problems = null;
+    this.problems = problems;
   }
 
   /** @return the number of rows in this table */
